+++ packages/docs/site/src/bin/makeEntries.js
@@ -1,13 +1,14 @@
 #!/usr/bin/env node -r esm
 
 import * as _ from 'lamb';
-import {concatValues, joinWith} from '@svizzle/utils';
+import {concatValues, joinWith, prepend} from '@svizzle/utils';
 
 import * as examples from '../routes/components/_examples';
 
 const makeEntries = _.pipe([
 	concatValues,
 	_.pluckKey('slug'),
+	_.mapWith(prepend('components/')),
 	joinWith(' ')
 ]);