+++ packages/tools/utils/src/array-[array-object].js
@@ -17,6 +17,8 @@ makeWithLatLng([1, 2]); // {lng: 1, lat: 2}
 makeWithLatLng([10, 20]); // {lng: 10, lat: 20}
  *
  * @version 0.2.0
+ * @see {@link module:@svizzle/utils/array-[array-object].makeWithValues|makeWithValues}
+ * @see {@link module:@svizzle/utils/array-[any-object].makeWith|makeWith}
  */
 export const makeWithKeys = keys => _.partial(_.make, [keys, _.__]);
 
@@ -33,5 +35,7 @@ makeWithTheseValues(["lng", "lat"]); // {lng: 1, lat: 2}
 makeWithTheseValues(["foo", "bar"]); // {foo: 1, bar: 2}
  *
  * @version 0.2.0
+ * @see {@link module:@svizzle/utils/array-[array-object].makeWithKeys|makeWithKeys}
+ * @see {@link module:@svizzle/utils/array-[any-object].makeWith|makeWith}
  */
 export const makeWithValues = values => _.partial(_.make, [_.__, values]);